downref exiting tree when setting a reference val
Downref a tree before replacing it in a reference. Also, don't need to walk the
list of references to set all kids to handle refs of refs. Normally these will
be split and the chain will be zero length. But even if they weren't, the kid
is the same and the tree in it needs to be replaced only once.